refactor(filtering): Move IItemFilter and MainItemFilter
