+++ app/src/main/java/com/example/ezvault/data/FilterRepository.java
@@ -3,8 +3,8 @@ package com.example.ezvault.data;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
-import com.example.ezvault.model.utils.filter.MainItemFilter;
 import com.example.ezvault.viewmodel.SortedItemListView;
+import com.example.ezvault.utils.MainItemFilter;
 
 import javax.annotation.Nullable;
 import javax.inject.Inject;
