+++ app/src/main/java/com/example/ezvault/model/utils/filter/ItemMakeFilter.java
@@ -1,6 +1,7 @@
 package com.example.ezvault.model.utils.filter;
 
 import com.example.ezvault.model.Item;
+import com.example.ezvault.utils.IItemFilter;
 
 /**
  * An item filter which ensures items conform to a specific make
