+++ app/src/main/java/com/example/ezvault/model/utils/filter/ItemTagFilter.java
@@ -2,6 +2,7 @@ package com.example.ezvault.model.utils.filter;
 
 import com.example.ezvault.model.Item;
 import com.example.ezvault.model.Tag;
+import com.example.ezvault.utils.IItemFilter;
 
 import java.util.List;
