+++ app/src/main/java/com/example/ezvault/utils/IItemFilter.java
@@ -1,4 +1,4 @@
-package com.example.ezvault.model.utils.filter;
+package com.example.ezvault.utils;
 
 import com.example.ezvault.model.Item;
